refactor(tests): split feeds.test.ts into list/detail/health files
480 行の feeds.test.ts を describe ブロック単位で 3 ファイルに分割。
articles/ ディレクトリの前例に倣い feeds/ ディレクトリを新設。

- feeds/list.test.ts  (~190 行): GET /api/feeds basic + フィルター
- feeds/detail.test.ts (~195 行): GET /api/feeds/:id
- feeds/health.test.ts (~195 行): GET /api/feeds/:id/health

Closes #364
